doing some handle bars stuff
