+++ package.json
@@ -1,5 +1,13 @@
 {
-    "scripts": {
-      "test": "jest"
-    }
+  "scripts": {
+    "test": "jest"
+  },
+  "dependencies": {
+    "dotenv": "^10.0.0",
+    "express": "^4.17.1",
+    "express-handlebars": "^5.3.3",
+    "handlebars": "^4.7.7",
+    "hbs": "^4.1.2",
+    "mysql2": "^2.3.0"
   }
+}
