+++ public/index.html
@@ -6,10 +6,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Gamio</title>
 </head>
+
+<script src="script.js"></script>
+
 <body>
     <h1>HELLO WORLD</h1>
 </body>
 
-<script src="script.js"></script>
 
 </html>
