+++ public/views/home.hbs
@@ -1,0 +1,9 @@
+<h1>Hello World from Handlebars</h1>
+
+<ul>
+    {{#each users}}
+            <li>{{this.name}}  {{this.address}}</li>
+    {{/each}}
+    <li>This is supposed to be a comment</li>
+    <li>This is supposed to be a comment</li>
+</ul>
