+++ public/views/layouts/main.hbs
@@ -1,0 +1,12 @@
+<html lang="en">
+<head>
+    <!-- <meta> tags> -->
+    <title>Book Face</title>
+</head>
+
+<body>
+    <div class="container">
+        {{{body}}}
+    </div>
+</body>
+</html>
