--- server.js
@@ -1,34 +1,0 @@
-const express = require("express");
-const mysql = require("mysql2");
-require("dotenv").config();
-
-const PORT = process.env.PORT || 8080;
-//create sql connection
-var con = mysql.createConnection({
-    host: "localhost",
-    user: "root",
-    password: process.env.password
-  });
-
-  con.connect(function(err) {
-    if (err) throw err;
-    console.log("Connected!");
-    con.query("CREATE DATABASE mydb", function (err, result) {
-      if (err) throw err;
-      console.log("Result: " + result);
-    });
-  });
-
-// Creating express app 
-const app = express();
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
-app.use(express.static("public"));
-
-app.listen(PORT, () => {
-console.log(
-    "==> 🌎  Listening on port %s. Visit http://localhost:%s/ in your browser.",
-    PORT,
-    PORT
-);
-});
